# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Family Finance Tracker",
    "version": "0.1.0"
  },
  "paths": {
...
    "/agent/chat": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Agent Chat",
        "description": "Run a conversational turn against the LangGraph coordinator.",
        "operationId": "agent_chat_agent_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTurnOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/resume": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Agent Resume",
        "description": "Resume a pending interrupt approval with approve or reject.",
        "operationId": "agent_resume_agent_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentResumeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTurnOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/thread/{thread_id}/state": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Agent Thread State",
        "description": "Inspect current checkpointer state for thread rehydration after reload or reconnect.",
        "operationId": "agent_thread_state_agent_thread__thread_id__state_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentThreadStateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/chat/stream": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Agent Chat Stream",
        "description": "Stream events (token/tool/artifact/interrupt/done) over SSE for long agent turns.",
        "operationId": "agent_chat_stream_agent_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentChatIn": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "message"
        ],
        "title": "AgentChatIn"
      },
      "AgentResumeIn": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "ops": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ops"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "decision"
        ],
        "title": "AgentResumeIn"
      },
      "AgentThreadStateOut": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageOut"
            },
            "type": "array",
            "title": "Messages"
          },
          "interrupted": {
            "type": "boolean",
            "title": "Interrupted"
          },
          "interrupt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterruptPayloadOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/TurnArtifactOut"
            },
            "type": "array",
            "title": "Artifacts",
            "default": []
          },
          "email_source_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmailSourceStatusOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "messages",
          "interrupted"
        ],
        "title": "AgentThreadStateOut"
      },
      "AgentTurnOut": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageOut"
            },
            "type": "array",
            "title": "Messages"
          },
          "interrupted": {
            "type": "boolean",
            "title": "Interrupted"
          },
          "interrupt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterruptPayloadOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/TurnArtifactOut"
            },
            "type": "array",
            "title": "Artifacts",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "messages",
          "interrupted"
        ],
        "title": "AgentTurnOut"
      },
...
      "ChatMessageOut": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "title": "Content"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tool_calls": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Calls"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessageOut"
      },
...
      "EmailSourceStatusOut": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "account_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Hint"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "available"
        ],
        "title": "EmailSourceStatusOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterruptPayloadOut": {
        "properties": {
          "ops": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Ops"
          },
          "preview": {
            "additionalProperties": true,
            "type": "object",
            "title": "Preview"
          },
          "preview_artifact_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Artifact Id"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          }
        },
        "type": "object",
        "required": [
          "ops",
          "preview"
        ],
        "title": "InterruptPayloadOut"
      },
...
      "TurnArtifactOut": {
        "properties": {
          "artifact_id": {
            "type": "integer",
            "title": "Artifact Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "digest": {
            "title": "Digest"
          }
        },
        "type": "object",
        "required": [
          "artifact_id",
          "kind",
          "title",
          "digest"
        ],
        "title": "TurnArtifactOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}